MoviesApp/Backend/RecommendationService/requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.21.1
gunicorn>=20.1.0
pydantic>=1.10.7

# Data processing and machine learning
pandas>=1.5.3
numpy>=1.24.2
scikit-learn>=1.2.2
scipy>=1.10.1

# Database connections
pyodbc>=4.0.35

# Environment and utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
httpx>=0.24.0
